requirements.txt
pinecone
python-dotenv
fastapi
uvicorn
google-genai
pydantic
groq
